+++ packages/core/src/plugin.ts
@@ -3,7 +3,8 @@ import * as React from 'react'
 import {
   StateDescriptor,
   StateDescriptorReturnType,
-  StateDescriptorSerializedType
+  StateDescriptorSerializedType,
+  StateDescriptorValueType
 } from './plugin-state'
 
 export type Plugin<
@@ -15,6 +16,9 @@ export type Plugin<
 export interface StatelessPlugin<Props extends Record<string, unknown> = {}> {
   Component: React.ComponentType<StatelessPluginEditorProps<Props>>
   onPaste?: (data: DataTransfer) => void | { state?: undefined }
+  title?: string
+  icon?: React.ComponentType
+  description?: string
 }
 
 export type StatelessPluginEditorProps<
@@ -34,6 +38,10 @@ export interface StatefulPlugin<
   onPaste?: (
     data: DataTransfer
   ) => void | { state?: StateDescriptorSerializedType<S> }
+  title?: string
+  icon?: React.ComponentType
+  description?: string
+  isEmpty?: (state: StateDescriptorValueType<S>) => boolean
   onKeyDown?: (e: KeyboardEvent) => boolean
   getFocusableChildren?: (
     state: StateDescriptorReturnType<S>