+++ packages/plugin-rows/src/row/clipboard.tsx
@@ -6,17 +6,28 @@ import {
   PluginState,
   getPlugins
 } from '@edtr-io/core'
-import { styled } from '@edtr-io/ui'
+import { styled } from '@edtr-io/editor-ui'
+import { ThemeProps } from '@edtr-io/ui'
+import { createRowPluginTheme } from '..'
 
 const ClipboardHeader = styled.div({
-  fontSize: '130%',
+  fontSize: 20,
   textAlign: 'center',
   marginBottom: '10px'
 })
 
-const Container = styled.div({
-  textAlign: 'center'
-})
+const Container = styled.div(
+  ({ name, ...props }: ThemeProps & { name: string }) => {
+    const theme = createRowPluginTheme(name, props.theme)
+    return {
+      textAlign: 'center',
+      backgroundColor: theme.menu.secondary.backgroundColor,
+      color: theme.menu.secondary.color,
+      padding: 15,
+      borderRadius: 5
+    }
+  }
+)
 
 const ButtonContainer = styled.div({
   marginTop: '10px',
@@ -44,11 +55,12 @@ const PreventMouseEvents = styled.div({
 
 export const Clipboard: React.FunctionComponent<{
   onClose: (pluginState: PluginState) => void
-}> = props => {
+  name: string
+}> = ({ name, ...props }) => {
   const store = React.useContext(EditorContext)
   const states = getClipboard(store.state)
   return (
-    <Container>
+    <Container name={name}>
       <ClipboardHeader> Zwischenablage </ClipboardHeader>
       <ButtonContainer>
         {states.length ? (