+++ packages/plugin-text/src/factory/index.tsx
@@ -1,7 +1,8 @@
 import { createTextEditor, SlateEditorAdditionalProps } from './editor'
 import { TextPluginOptions } from './types'
 import { StateType, StatefulPlugin } from '@edtr-io/core'
-import { ValueJSON } from 'slate'
+import { Value, ValueJSON } from 'slate'
+import { StateDescriptorValueType } from '@edtr-io/core/src/plugin-state'
 
 export const defaultNode = 'paragraph'
 
@@ -35,6 +36,10 @@ export const createTextPlugin = (
     state: textState,
     onKeyDown() {
       return false
+    },
+    isEmpty: (state: StateDescriptorValueType<typeof textState>) => {
+      const value = Value.fromJSON(state)
+      return value.document.text === ''
     }
   }
 }