+++ CHANGELOG.md
@@ -22,6 +22,10 @@ All notable changes to this project will be documented in this file.
 - **plugin-text**. Hide placeholder in render mode
 - **plugin-text**. Hovering Menu for text formatting
 - **ui** / **editor-ui** / **renderer-ui**. Revised theming workflow
+- **plugin-rows**. Improved styling of add menu and controls
+- **plugin-rows**. Added more theming options
+- **core**. Added `isEmpty(state: State, id: string)` to check if a plugin state is empty
+- **core**. Added optional function `isEmpty: (state: StateDescriptorValueType<S>) => boolean` to plugin definition
 
 ### Fixed